完成异步队列

+++ src/main/java/me/dack/wenda/controller/QuestionController.java
@@ -11,8 +11,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import me.dack.wenda.model.Errcode;
+import me.dack.wenda.model.HostHolder;
 import me.dack.wenda.model.Question;
 import me.dack.wenda.model.Result;
+import me.dack.wenda.model.User;
 import me.dack.wenda.service.QuestionService;
 
 @Controller
@@ -23,6 +26,8 @@ public class QuestionController {
 	
 	@Autowired
 	private QuestionService questionService;
+	@Autowired
+	private HostHolder hostHolder;
 	
 	
 	@RequestMapping("/addQuestion")
@@ -33,19 +38,20 @@ public class QuestionController {
 		question.setTitle(title);
 		question.setContent(content);
 		question.setCreateTime(new Date());
-		//
 		question.setCommentCount(0);
 		question.setStatus(0);
-		question.setUserId(1);
+		
+		User user = hostHolder.getUser();
+		question.setUserId(user.getId());
 		try{
 			if(questionService.addQuestion(question) > 0){
-				Result result = new Result(true, "添加成功");
+				Result result = new Result(Errcode.Null, "添加成功");
 				return result;
 			}
 		}catch (Exception e) {
 			logger.error("添加问题失败"+e.getMessage());
 		}
-		Result result = new Result(false, "添加失败");
+		Result result = new Result(Errcode.Error, "添加失败");
 		return result;
 	}
 	
@@ -55,12 +61,13 @@ public class QuestionController {
 			@RequestParam("limit")int limit){
 		try{
 			List<Question> latestQuestions = questionService.getLatestQuestions(userId, offset, limit);
-			Result result = new Result(true, "查找失败");
+			Result result = new Result(Errcode.Null, "查找成功");
 			result.setRes(latestQuestions);
+			return result;
 		}catch (Exception e) {
 			logger.error("查看最新问题失败"+e.getMessage());
 		}
-		Result result = new Result(false,"查找失败");
+		Result result = new Result(Errcode.Error,"查找失败");
 		return result;
 	}
 	
@@ -70,13 +77,13 @@ public class QuestionController {
 			@RequestParam("content")String content){
 		try{
 			if(questionService.updateQuestion(id, title, content) > 0){
-				Result result = new Result(true, "修改成功");
+				Result result = new Result(Errcode.Null, "修改成功");
 				return result;
 			}
 		}catch (Exception e) {
 			logger.error("修改问题内容失败"+e.getMessage());
 		}
-		Result result = new Result(false,"修改失败");
+		Result result = new Result(Errcode.Error,"修改失败");
 		return result;
 	}
 	
@@ -84,14 +91,19 @@ public class QuestionController {
 	@ResponseBody
 	public Result deleteQuestion(@RequestParam("id")int id){
 		try{
+			Question question = questionService.getQuestionById(id);
+			if (question.getUserId() != hostHolder.getUser().getId()) {
+				Result result = new Result(Errcode.Error,"异常请求");
+				return result;
+			}
 			if(questionService.deleteQuestion(id) > 0){
-				Result result = new Result(true, "删除成功");
+				Result result = new Result(Errcode.Null, "删除成功");
 				return result;
 			}
 		}catch (Exception e) {
 			logger.error("删除问题内容失败"+e.getMessage());
 		}
-		Result result = new Result(false,"删除失败");
+		Result result = new Result(Errcode.Error,"删除失败");
 		return result;
 	}
 }
